+++ micropico.py
@@ -1,5 +1,7 @@
 from src.keypad.keypad import Keypad
-from src.led.led import Led
+from src.led.pwm_led import PWMLed
+from src.led.digital_led import DigitalLed
+from src.rgbled.rgb_led import RGBLed
 from src.photoresistor.photoresistor import Photoresistor
 from src.ultrasonic.ultrasonic import Ultrasonic
 from src.potentiometer.potentiometer import Potentiometer
