--- src/interfaces/output_device.py
@@ -1,24 +1,0 @@
-class OutputDevice:
-    def __init__(self):
-        pass
-    
-    @property
-    def value(self):
-        pass
-    
-    @property.setter
-    def value(self):
-        pass
-    
-    @property
-    def working(self):
-        pass
-    
-    def on(self):
-        pass
-    
-    def off(self):
-        pass
-    
-    def toggle(self):
-        pass
