+++ src/rgbled/rgb_led.py
@@ -1,0 +1,10 @@
+from src.interfaces.base_led import BaseLed
+
+
+class RGBLed(BaseLed):
+    """
+    Represents rgb diode or a rgb strip.
+    """
+    def __init__(self, red: int, green: int, blue: int):
+        pass
+
